+++ alh_pdf_view_platform_interface/lib/alh_pdf_view_platform_interface.dart
@@ -3,7 +3,6 @@ export 'package:alh_pdf_view_platform_interface/src/platform_interface/alh_pdf_v
 export 'package:alh_pdf_view_platform_interface/src/types/alh_pdf_view_creation_params.dart';
 export 'package:alh_pdf_view_platform_interface/src/types/fit_policy.dart';
 export 'package:alh_pdf_view_platform_interface/src/types/map_event/page_changed_object.dart';
-export 'package:alh_pdf_view_platform_interface/src/types/map_event/page_error_object.dart';
 export 'package:alh_pdf_view_platform_interface/src/types/typedefs.dart';
 export 'package:alh_pdf_view_platform_interface/src/types/unknown_view_id_error.dart';
 export 'package:alh_pdf_view_platform_interface/src/types/widget_configuration.dart';