+++ alh_pdf_view_platform_interface/lib/src/events/map_event.dart
@@ -1,5 +1,4 @@
 import 'package:alh_pdf_view_platform_interface/src/types/map_event/page_changed_object.dart';
-import 'package:alh_pdf_view_platform_interface/src/types/map_event/page_error_object.dart';
 
 /// This class is used to notify method calls from the native part.
 class MapEvent<T> {
@@ -20,10 +19,6 @@ class OnPageChangedEvent extends MapEvent<PageChangedObject> {
 
 class OnErrorEvent extends MapEvent<String> {
   OnErrorEvent(super.viewId, super.error);
-}
-
-class OnPageErrorEvent extends MapEvent<PageErrorObject> {
-  OnPageErrorEvent(super.viewId, super.pageErrorObject);
 }
 
 class OnZoomChangedEvent extends MapEvent<double> {