+++ alh_pdf_view_platform_interface/lib/src/types/map_event/page_error_object.dart
@@ -1,9 +1,0 @@
-class PageErrorObject {
-  final int page;
-  final String error;
-
-  const PageErrorObject({
-    required this.page,
-    required this.error,
-  });
-}